requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Naked==0.1.31
pycryptodome==3.10.1
requests==2.25.1
shellescape==3.8.1
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.1
